+++ ost_visualizer/presentation/handlers/export_handler.py
@@ -159,6 +159,8 @@ class ExportHandler:
                     config.grayscale_enabled,
                     caption_settings=caption_settings,
                     elevation_callouts_enabled=config.pdf_elevation_callouts_enabled,
+                    elevation_callout_settings=config.elevation_callout_settings(),
+                    elevation_callout_color=config.pdf_elevation_callout_color,
                     page_area_selections=(self.project_data.get_page_area_selections()),
                     bid_annotations=bid_annotations,
                     on_progress=_progress_callback(reporter),